fix the modification bug (when img doesn't exist)

+++ controllers/Post.js
@@ -195,12 +195,13 @@ exports.modifyPost = (req, res, next) => {
             if (post.userId === req.token.userId || req.token.isAdmin) {
                 let postObject;
                 if (req.file) {
-                    const filename = post.imageUrl.split('/images/')[1];
-                    fs.unlinkSync(`images/${filename}`);
-
+                    if (post.imageUrl) {
+                        const filename = post.imageUrl.split('/images/')[1];
+                        fs.unlinkSync(`images/${filename}`);
+                    }
+                    
                     postObject = {
                         ...req.body,
-                        // userId: req.token.userId,
                         imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
                     }
                 } else {
